basic unified circuit impl w/o tests
