+++ circuit/Cargo.toml
@@ -1,12 +1,15 @@
 [package]
 version = "0.1.0"
-name = "zktron"
+name = "untron-circuit"
 edition = "2021"
 
 [dependencies]
 alloy-sol-types = "0.7.2"
 sp1-zkvm = "1.0.1"
+hex-literal = "0.4.1"
 sha2 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", branch = "patch-sha2-v0.10.8" }
 ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", features = ["verifying"], branch = "patch-ecdsa-v0.16.9" }
-hex-literal = "0.4.1"
 k256 = { version = "0.13", default-features = false, features = ["ecdsa"] }
+bincode = "1.3.3"
+serde = { version = "1.0.208", features = ["derive"] }
+sha3 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha3", branch = "patch-sha3-v0.10.8" }
