--- circuits/test/walkthrough/rust-toolchain
@@ -1,3 +1,0 @@
-[toolchain]
-channel = "nightly-2024-04-17"
-components = ["llvm-tools", "rustc-dev"]
