--- circuits/test/zktron/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-version = "0.1.0"
-name = "zktron-test"
-edition = "2021"
-
-[dependencies]
-sp1-sdk = { git = "https://github.com/succinctlabs/sp1.git", tag = "v1.0.7-testnet" }
-sp1-core = { git = "https://github.com/succinctlabs/sp1.git", tag = "v1.0.7-testnet" }
-k256 = "0.13.3"
-serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
-serde = { version = "1.0", default-features = false, features = ["derive"] }
-clap = { version = "4.0", features = ["derive", "env"] }
-tracing = "0.1.40"
-alloy-sol-types = "0.7.2"
-hex = "0.4.3"
-rand = "0.8.5"
-sha2 = "0.10.8"
-tokio = { version = "1.39.2", features = ["full"] }
-tonic = { version = "0.12.1", features = ["prost"] }
-prost = "0.13.1"
-prost-types = "0.13.1"
-
-[build-dependencies]
-sp1-helper = { git = "https://github.com/succinctlabs/sp1.git", tag = "v1.0.7-testnet" }
-tonic-build = "0.12.1"
