--- circuits/test/zktron/src/srs.json
@@ -1,29 +1,0 @@
-[
-    "e011f700743efb9368e1944182be7a49dc1552aa630cc40032ae8f73cd538dffce12577d142ea3d97185f66ad98c1d3313cbe7701e491aa3e694fbc576a3c4a1",
-    "9574859b9b7164ce8461f309ac1c376b6fb432dfdad6835a8b3a89c1081c09686d63802cace65f103c6ec6caeee841d384130ba3895c62a4abe3864d623cadc2",
-    "3c1926539d8e26002232c7747fd755dc329729fa06d9bb33036df1642c2841a3172fa742122cd669eb060a34837f57892b2dc6845668757655dfdbd9fae1607c",
-    "a7a28698357fa74e6cd4f58fa282944ceae178e21d7d8b3e86298db62b83dad9845b50918c7a16b66539468ef2f4d3d81ab5a683cfeb533dd026bacfa7b987d3",
-    "96c29979399c504d7dda7f8eee96a64d564fba2cf79453bb95dbddd066f2d79cb4273ad7faefb745fbea06ee8e27c391f04eae0f65639faa3cce556b4749208e",
-    "1e70ffc9fc88e7f3ebad47ace51e25110afbbf43defe050198a65e39ef26bd32abfdc42fc748310e15bcb012631db9c91bfb287a71badf14bc12afe225c15682",
-    "567884306c368b1bdeab87b4633ba0110bf447bf031e20da1391ef5721f46d027ff5906de0a377410edcc3a85561b5c53c364bedbbdac7c10c7bfdb0b7b7dc0d",
-    "7f4682720ee832013a384b680d555ddfafbfc084420a2712b3ea4cd81879e0814dcf29ade8b73f16a74477358d7aa601dd7600e38019d70883124e36f960cdd2",
-    "daad914514b8a4edfe67b839933e516197df4131902a0cf01568cd41a68df96beed1f974c98648e6a3b7b8dd9601b7a76ab3835849f66334e246c95c787e0555",
-    "20a1cc3d8c201e6793b5f58f562d127dcbe92da89558b0b574bd5ebfd7dd30aaf1a7764f0ea5345616198d94db436c69ea419430c6727978ed500f85682feb01",
-    "52f9adf2e2ebeae93dbbe7da8eff2f8fe8367c5a968ab907985af8435db394a8fd3795d34eb397abf7a5128b0092a32ff1dd24d72e033e77ffd7bc67bf3ed892",
-    "202e4920a434080c59e7202b1feb84e37b67d63111ae055223aa34337893818d2d6df95c5bf4502d219200ff4b256fb87368e68fd6b73f407de8a3d4caf87168",
-    "6318cadcc796974316207aa905f8bb21df714e97d5e393c00ae7acce9bc49f36b60c3339f0814a64bbad1738f7946cfb12f5188e7216bc6c038479b8ff1db816",
-    "cde030342f75f6061294c12ec0e8f4723f74e7d4821e21451c80e9a2c601ecbc43c919fac4e450d8a691c9777f0d8a775da707468d955edeb65ae7f6ac011387",
-    "67943aad01bd2375d46c19cf9e8ba04d46490513eb9a0091b3767faf2a13faab4a0599602bade1f8de3cc9d35ecc214c83a3672d2ddbd9677eab3b42faba2700",
-    "12b50d6895e6010f0f7fb4e6eba00fb4eca46229649b60520bc09f8bb3b9dc26d66ab4752a2f3bd6a5e517b6a173a0a6f1cbe4867a0195d2bfeb9f823817a9e0",
-    "f5163b6ad0ddc2fd0918025fac1570a99a66fa2d61ef8a6f8e78e2b3ee6394d124fc5635161ade29471690ec7d975c2cede5621756f85c954ac365d73f0916b4",
-    "eb3efef18c474fc6f51a2992a51a0e2cf7e26e0bb744fb4685d604d5bd5f09277df5cbb9d603b5fa73694a8f05c6de042c31d172057b999e0e5be71863cec961",
-    "f37e61068c203bb383ad838be0c4f17a4b5b8aed610e23214b16ce73e645847c37a44e74e8ec5bad7cdcb0307d5d818d4f182238b9712fe3fb97f66afe7ebdc8",
-    "2dc9dec7e73146f61dd3635e37f1d03eef05905f4b11152955f01dc497e1c10030234292adc2fe0627cfaf14be4d3a28dc6423f6b9f5a23cfc6c590e09890802",
-    "6058dce81b8eb6ac7c45d3f312feb0eeb838c0f18687c58f8b76ec3df07b04055533e68f31221f17b4549e824a7dc91cf131ea53bab30537f4c498b44e82345d",
-    "dd43d952d0956e693dd5ae4669115a5ab7021b014aedf5e47f12c268364a9f9e0716ed76da72f8fa9e79336de7d45902c1a5dadd337ad9ba0f9924cebb984ba0",
-    "e7a94c9ec18a865874eea88117240166f69d44da1cd73e0b4391956845446b655fc91094da89cb943d252d9a85aeaabf5043eeaa2286132889cad165dd306332",
-    "c6bb1785110bdc360794f9a05420d09ac646cfec5ff8d4c7e37de03bb3729fe29ecc79b196dec61a5b33eb793aee854a39f143d9096f7c9594f60a99d59b7913",
-    "717a7fb111240f13077f262146a63f8f94dbeea4c58354e19d1b9d566c3ebb28c7685b8a9c61192e50f8664aef872a056d85b9b27d52f8cbefadea35dd182613",
-    "a61e16d92d7bb807f687e0d67e6f1da5bcddfe7bcd1f6e90a910c220fb29ec46ca568916b95af36c06f80423511b34d2a6542c82035432395ec7a3084a9de2ff",
-    "67de4836526c8f2097aa7e273b1982bc5abfbcb49a7694bc910ae74abcff8d545b5c74f078bf992cf4c470610c78e81fac8adfe46c4eabd2d1056852539cb049"
-]
