--- circuits/walkthrough/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-version = "0.1.0"
-name = "walkthrough"
-edition = "2021"
-
-[dependencies]
-alloy-sol-types = "0.7.2"
-sp1-zkvm = "1.0.1"
-hex-literal = "0.4.1"
-zktron = { path = "../zktron" }
