--- circuits/zktron/README.md
@@ -1,11 +1,0 @@
-# Untron.finance: ZK light client for Tron Network
-
-This directory contains the ZK light client for Tron Network ("ZKTron") written in SP1 zkVM. ZKTron verifies and parses block headers and returns data necessary for Untron bridge. It's capable of ZK proving large chains of blocks in Tron in a short time, making it useful in the hours of low activity in the bridge.
-
-## Making a reproducible build
-
-`cargo prove build --docker --tag v1.0.1`
-
-## Checksum
-
-`TODO`
